Add exec as a subcommand
