+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/sasanrose/gbench/cmd"
+
+func main() {
+	cmd.Execute()
+}
